"Updated Error.mjs: changed argument of TypeError in TypesCheck function from 'Length' to 'Type'"
This commit message describes the change in the code diff for the file Error.mjs. Specifically, it mentions that the argument of the TypeError constructor in the TypesCheck function has been updated from 'Length' to 'Type'. This commit message is concise and accurately describes the changes made in the code diff.